refactor: extract shared fidelity bond sweep flow from renew and move dialogs
